Refactor functions and add abac
